moved bot to lib
