--- stored_tweets.yml
@@ -1,23 +1,0 @@
----
-b•gall: New Orleans, LA
-bicentennial baby 76: United States
-JSBot: Raspberry PI
-Kayyzouuツ: France
-Moorestown NJ Patch: Moorestown, NJ
-Snafets: sc, twitch, tik tok, instagram
-Angel: Taurus | Virgo | Virgo
-DonnaChaffins: Knoxville, TN
-"\U0001F4AB": In my bag
-walter medina: CARACAS- VENEZUELA
-"ذكريات \U0001F5A4": للاعلانات ونظام الرتويت DM
-しゃんてぃ@よしこのかみのけたべたいbot: ﾒｩﾄﾓﾃﾗﾎｼｽ
-"-スカイドローズ-": Lille, France
-NearSunrise: Scotland
-arclight: Austin, TX
-"\U0001F98B Elemental Essence \U0001F98B": Maine, USA
-"Paul Mwiu \U0001F1F0\U0001F1EA\U0001F30D": Nairobi
-"!rungu": The Cradle of Mankind
-Python Hunt Bot: 'Internet '
-xael bot: Earth
-Profitrool: Gard, Languedoc-Roussillon
-"SafariSara \U0001F30E\U0001F1EA\U0001F1FA #FBPE": Here - There - Everywhere
